[execution] New TU `execute_block_header`
This patch hoists the functions `execute_block_header` and
`set_beacon_root` into a separate translation unit. The reason for this
change is to enable those functions to be reused by the zkVM target.
