+++ category/execution/ethereum/execute_block.hpp
@@ -47,9 +47,6 @@ namespace fiber
 } // namespace fiber
 
 template <Traits traits>
-void execute_block_header(BlockState &, BlockHeader const &);
-
-template <Traits traits>
 Result<std::vector<Receipt>> execute_block_transactions(
     Chain const &, BlockHeader const &, std::span<Transaction const>,
     std::span<Address const> senders,
